+++ linen_examples/imagenet/models.py
@@ -86,13 +86,12 @@ class ResNet(nn.Module):
   num_filters: int = 64
   dtype: Any = jnp.float32
   act: Callable = nn.relu
-  train: bool = True
 
   @nn.compact
-  def __call__(self, x):
+  def __call__(self, x, train: bool = True):
     conv = partial(nn.Conv, use_bias=False, dtype=self.dtype)
     norm = partial(nn.BatchNorm,
-                   use_running_average=not self.train,
+                   use_running_average=not train,
                    momentum=0.9,
                    epsilon=1e-5,
                    dtype=self.dtype)