+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 vNext
 ----
  (Please add entries here with your changes for the next version)
+ - `flax.core.apply` and linen `Module.apply` will now only return the variables
+   collections that were specified as mutable.
  - ...
 
 v0.2.2